+++ h2integrate/core/finances.py
@@ -1,9 +1,16 @@
+from pathlib import Path
+
 import numpy as np
 import ProFAST  # system financial model
 import openmdao.api as om
 import numpy_financial as npf
 
-from h2integrate.core.utilities import check_plant_config_and_profast_params
+from h2integrate.core.utilities import (
+    dict_to_yaml_formatting,
+    check_plant_config_and_profast_params,
+)
+from h2integrate.core.inputs.validation import write_yaml
+from h2integrate.tools.profast_reverse_tools import convert_pf_to_dict
 
 
 class AdjustedCapexOpexComp(om.ExplicitComponent):
@@ -332,6 +339,18 @@ class ProFastComp(om.ExplicitComponent):
         # ------------------------------------ solve and post-process -----------------------------
 
         sol = pf.solve_price()
+
+        # Check whether to export profast object to .yaml file
+        if self.options["plant_config"]["finance_parameters"].get("save_profast_to_file", False):
+            output_dir = self.options["driver_config"]["general"]["folder_output"]
+            fdesc = self.options["plant_config"]["finance_parameters"]["profast_output_description"]
+            fname = f"{fdesc}_{self.options['commodity_type']}.yaml"
+            fpath = Path(output_dir) / fname
+            output_dir = Path(output_dir)
+            output_dir.mkdir(parents=True, exist_ok=True)
+            d = convert_pf_to_dict(pf)
+            d = dict_to_yaml_formatting(d)
+            write_yaml(d, fpath)
 
         # Only hydrogen supported in the very short term
         if self.options["commodity_type"] == "hydrogen":